security warning

+++ didJavaScriptFunctionRun/scripts/delete_random_file.php
@@ -4,18 +4,23 @@
   
   // Include cleaning function.
   include("clean_random_character_function.php");
+  // include check user agent.
+  include("check_header_did_javascript_function_run.php");
+  check_header_did_javascript_function_run();
   
-  $out_uri = $_SERVER["REQUEST_URI"];
-  $out_query = substr($out_uri, strrpos($out_uri, "?")+1);
-  $out_query = urldecode($out_query);
-  /*$out_query = trim($out_query);
-  $out_query = htmlspecialchars($out_query);
-  $out_query = preg_replace('/[.\/=$]/', "", $out_query);*/
-  $out_query = clean_random_character_function($out_query);
-  
-  $out_q_file = "../tmp/$out_query";
-  
-  if (file_exists($out_q_file . "/file.txt")) {
-   shell_exec("rm -rf $out_q_file");
+  if ($user_agent_did_javascript_function_run == 0) {
+   $out_uri = $_SERVER["REQUEST_URI"];
+   $out_query = substr($out_uri, strrpos($out_uri, "?")+1);
+   $out_query = urldecode($out_query);
+   /*$out_query = trim($out_query);
+   $out_query = htmlspecialchars($out_query);
+   $out_query = preg_replace('/[.\/=$]/', "", $out_query);*/
+   $out_query = clean_random_character_function($out_query);
+
+   $out_q_file = "../tmp/$out_query";
+
+   if (file_exists($out_q_file . "/file.txt")) {
+    shell_exec("rm -rf $out_q_file");
+   }
   }
 ?>
